fix(richtext-lexical): editor throws an error if OrderedList is registered but not UnorderedList or CheckList (#14149)
The problem was that the nodes weren't being added.

The logic to avoid adding the same node or plugin multiple times was
repeated in several places. It was also difficult to understand and easy
to make mistakes.

I extracted the logic into a new `shouldRegisterList` utility and
corrected the behavior where necessary.

Fixes #14148