fix(plugin-cloud-storage): file sizes use correct mime type (#11633)
### What?

Fixes #7575

### Why?

The bug was caused by using the upload file mimeType instead of the
resized file mime types.

Often, the resized file will have a different mime type (ex: converting
all files to jpg) so this is incorrect.

### How?

Use the resizedFile mime type with the upload file mimeType as a backup

I'm unsure when/why the resized files wouldn't have a mime type, but the
base file mime type is a sensible backup option.