test: deflakes localization e2e suite (#14639)
The localization e2e suite would use the `runCopy` helper function which
copies data from 1 locale to another locale and then navigates to that
locale.

Tests would fail intermittently because they did not wait for that route
change to happen. Scenarios would arise where the request took longer
than the next `changeLocale` call which changes the locale and navigates
to that locale.

So it looked like:
- `runCopy` is called **but has not finished**
- `changeLocale` runs and navigates user to new locale
- `runCopy` finishes and navigates to the locale it copied to