fix(ui): autosave not queued while background process is in flight (#14805)
When editing fields while an autosave is in flight, subsequent autosave
events are not queued as expected. This results in inconsistency between
what appears on the screen and what is persisted to the database, due to
race conditions in background processing.

For example:

https://github.com/user-attachments/assets/010f846c-ed07-4594-883e-3c60114e9f9e

This is especially noticeable in virtual fields, `useAsTitle`, and Live
Preview, where the displayed values are a direct representation of the
database, not local to form state.

The problem is related to how we manage `modified` form context. After
successful submissions, we set this flag to false. This means if the
user had made underlying changes to the form while a previous submission
was processing, e.g. an autosave, the modified state gets wrongfully
cleared and the next autosave bails out.

Potentially related: https://github.com/payloadcms/payload/issues/12691