fix(ui): component AddNewRelation saves relationTo as first collection instead of selected collection (#14836)
### What?

Currently when you use the create button in a relationship field that
has more than one collection in its `relationTo` and a collection that
isn't first in the list is selected, the selected value in the field
will always be pointing to the first collection in the list, causing
incorrect results or validation errors.

### Why?

The selected value in the field should be applied to the collection the
user has selected, not just the first collection in the `relationTo`
array.

### How?

Use the selected collection in the `onChange` handler instead of
`relatedCollections[0]`.

Fixes: #14728

---------

Co-authored-by: Patrik Kozak <35232443+PatrikKozak@users.noreply.github.com>