docs: update collection access control reference link on collection config page (#14905)
The link despite written as "Collection Access Control", currently point
to the general Access Control page and not the dedicated page for
Collection Access Control. On the same paragraph there's also a
reference to the Access Control page and makes it confusing.
This replace the link to point to the dedicated Collection Access
Control page.